Switch from using 2-way stream to 1-way stream + QSharedMemory
Instead of communicating both directions on a single pipe, switch
to using the pipe only for Python->C++, and then have C++ write
to a shared memory location.  Once both detach from the shared
memory, it will automatically be deleted, so there's no leak.

This should finally resolve the server getting stuck in a state
where the pipe is out of sync with request/read order.